+++ Common.cs
@@ -3,6 +3,7 @@ using System.Runtime.InteropServices;
 using Dalamud.Game.Chat.SeStringHandling;
 using Dalamud.Plugin;
 using FFXIVClientStructs;
+using SimpleTweaksPlugin.GameStructs;
 
 namespace SimpleTweaksPlugin {
     internal class Common {
@@ -15,15 +16,49 @@ namespace SimpleTweaksPlugin {
 
         private static GameAlloc gameAlloc;
         private static GetGameAllocator getGameAllocator;
-        
 
+        private delegate IntPtr GetInventoryContainer(IntPtr inventoryManager, int inventoryId);
+        private unsafe delegate InventoryItem* GetContainerSlot(IntPtr inventoryContainer, int slotId);
+
+        private static GetInventoryContainer getInventoryContainer;
+        private static GetContainerSlot getContainerSlot;
+
+        private static IntPtr inventoryManager;
+
+        public static IntPtr PlayerStaticAddress { get; private set; }
 
         public Common(DalamudPluginInterface pluginInterface) {
             this.pluginInterface = pluginInterface;
             var gameAllocPtr = pluginInterface.TargetModuleScanner.ScanText("E8 ?? ?? ?? ?? 45 8D 67 23");
             var getGameAllocatorPtr = pluginInterface.TargetModuleScanner.ScanText("E8 ?? ?? ?? ?? 8B 75 08");
+
+            inventoryManager = pluginInterface.TargetModuleScanner.GetStaticAddressFromSig("BA ?? ?? ?? ?? 48 8D 0D ?? ?? ?? ?? E8 ?? ?? ?? ?? 48 8B F8 48 85 C0");
+            var getInventoryContainerPtr = pluginInterface.TargetModuleScanner.ScanText("E8 ?? ?? ?? ?? 8B 55 BB");
+            var getContainerSlotPtr = pluginInterface.TargetModuleScanner.ScanText("E8 ?? ?? ?? ?? 8B 5B 0C");
+
+            PlayerStaticAddress = pluginInterface.TargetModuleScanner.GetStaticAddressFromSig("8B D7 48 8D 0D ?? ?? ?? ?? E8 ?? ?? ?? ?? 0F B7 E8");
+
             gameAlloc = Marshal.GetDelegateForFunctionPointer<GameAlloc>(gameAllocPtr);
             getGameAllocator = Marshal.GetDelegateForFunctionPointer<GetGameAllocator>(getGameAllocatorPtr);
+
+            getInventoryContainer = Marshal.GetDelegateForFunctionPointer<GetInventoryContainer>(getInventoryContainerPtr);
+            getContainerSlot = Marshal.GetDelegateForFunctionPointer<GetContainerSlot>(getContainerSlotPtr);
+        }
+
+        public static IntPtr GetContainer(int containerId) {
+            if (inventoryManager == IntPtr.Zero) return IntPtr.Zero;
+            return getInventoryContainer(inventoryManager, containerId);
+        }
+
+        public static unsafe InventoryItem* GetContainerItem(IntPtr container, int slot) {
+            if (container == IntPtr.Zero) return null;
+            return getContainerSlot(container, slot);
+        }
+
+        public static unsafe InventoryItem* GetInventoryItem(int inventoryId, int slotId) {
+            if (inventoryManager == IntPtr.Zero) return null;
+            var container = getInventoryContainer(inventoryManager, inventoryId);
+            return container == IntPtr.Zero ? null : getContainerSlot(container, slotId);
         }
 
         public static IntPtr Alloc(ulong size) {