+++ UiHelper.cs
@@ -32,9 +32,9 @@ namespace SimpleTweaksPlugin {
             SetText(textNode, seStr);
         }
 
-        public static void SetSize(AtkResNode* node, ushort? width, ushort? height) {
-            if (width != null) node->Width = width.Value;
-            if (height != null) node->Height = height.Value;
+        public static void SetSize(AtkResNode* node, int? width, int? height) {
+            if (width != null && width >= ushort.MinValue && width <= ushort.MaxValue) node->Width = (ushort) width.Value;
+            if (height != null && height >= ushort.MinValue && height <= ushort.MaxValue) node->Height = (ushort) height.Value;
             node->Flags_2 |= 0x1;
         }
 
@@ -73,6 +73,52 @@ namespace SimpleTweaksPlugin {
             SetPosition(n, width - 61, 8); // Help Button
 
             windowNode->AtkResNode.Flags_2 |= 0x1;
+        }
+
+        public static void ExpandNodeList(AtkComponentNode* componentNode, ushort addSize) {
+            var newNodeList = ExpandNodeList(componentNode->Component->ULDData.NodeList, componentNode->Component->ULDData.NodeListCount, (ushort) (componentNode->Component->ULDData.NodeListCount + addSize));
+            componentNode->Component->ULDData.NodeList = newNodeList;
+        }
+
+        public static void ExpandNodeList(AtkUnitBase* atkUnitBase, ushort addSize) {
+            var newNodeList = ExpandNodeList(atkUnitBase->ULDData.NodeList, atkUnitBase->ULDData.NodeListCount, (ushort)(atkUnitBase->ULDData.NodeListCount + addSize));
+            atkUnitBase->ULDData.NodeList = newNodeList;
+        }
+
+        private static AtkResNode** ExpandNodeList(AtkResNode** originalList, ushort originalSize, ushort newSize = 0) {
+            if (newSize <= originalSize) newSize = (ushort)(originalSize + 1);
+            var oldListPtr = new IntPtr(originalList);
+            var newListPtr = Alloc((ulong)((newSize + 1) * 8));
+            var clone = new IntPtr[originalSize];
+            Marshal.Copy(oldListPtr, clone, 0, originalSize);
+            Marshal.Copy(clone, 0, newListPtr, originalSize);
+            return (AtkResNode**)(newListPtr);
+        }
+
+        public static AtkResNode* CloneNode(AtkResNode* original) {
+            var size = original->Type switch
+            {
+                NodeType.Res => sizeof(AtkResNode),
+                NodeType.Image => sizeof(AtkImageNode),
+                NodeType.Text => sizeof(AtkTextNode),
+                NodeType.NineGrid => sizeof(AtkNineGridNode),
+                NodeType.Counter => sizeof(AtkCounterNode),
+                NodeType.Collision => sizeof(AtkCollisionNode),
+                _ => throw new Exception($"Unsupported Type: {original->Type}")
+            };
+
+            var allocation = Alloc((ulong)size);
+            var bytes = new byte[size];
+            Marshal.Copy(new IntPtr(original), bytes, 0, bytes.Length);
+            Marshal.Copy(bytes, 0, allocation, bytes.Length);
+
+            var newNode = (AtkResNode*)allocation;
+            newNode->ParentNode = null;
+            newNode->ChildNode = null;
+            newNode->ChildCount = 0;
+            newNode->PrevSiblingNode = null;
+            newNode->NextSiblingNode = null;
+            return newNode;
         }
 
     }